requirements.txt
certifi==2024.8.30
cftime==1.6.4
charset-normalizer==3.3.2
greenlet==3.1.1
idna==3.10
netCDF4==1.7.1.post2
numpy==2.0.2
pip-system-certs==4.0
psycopg2-binary==2.9.9
requests==2.32.3
typing_extensions==4.12.2
urllib3==1.26.6
wrapt==1.16.0
